Following are part of this commit +++++++++++++++++++++++++++++++++ * Windows CNI Support 	(1) Support to use --network-plugin=cni 	(2) Handled platform requirement of calling CNI ADD for all the containers. 	(2.1) For POD Infra container, netNs has to be empty 	(2.2) For all other containers, sharing the network namespace of POD container, 	      should pass netNS name as "container:<Pod Infra Container Id>", same as the               NetworkMode of the current container 	(2.3) The Windows CNI plugin has to handle this to call into Platform.               Sample Windows CNI Plugin code to be shared soon. * Sandbox support for Windows 	(1) Sandbox support for Windows. Works only with Docker runtime. 	(2) Retained CONTAINER_NETWORK as a backward compatibilty flag, 	    to not break existing deployments using it. 	(3) Works only with CNI plugin enabled.
(*) Sample Script to start Kubelet and KubeProxy
(*) Changes to reinvoke CNI ADD for every new container created. This is hooked up with PodStatus,
    but would be ideal to move it outside of this, once we have CNI GET support

From @madhanrm, kubernetes/kubernetes PR #51063
